# Add zlib-ng library, optimized for ARM64 