… src/libraries/Native/Windows/CMakeLists.txt
… src/libraries/Native/Windows/System.IO.Compression.Native/CMakeLists.txt
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/CMakeLists.txt
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/LICENSE.md
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/README.md
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/compare258.c
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/compress.c
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/configure
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/crc32.c
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/crc32_comb.c
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/crc32_comb_tbl.h
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/crc32_p.h
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/crc32_tbl.h
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/deflate.c
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/deflate.h
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/deflate_fast.c
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/deflate_medium.c
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/deflate_p.h
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/deflate_quick.c
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/deflate_slow.c
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/fallback_builtins.h
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/functable.c
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/functable.h
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/gzguts.h
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/gzlib.c
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/gzread.c
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/gzwrite.c
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/infback.c
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/inffast.c
… src/libraries/Native/Windows/System.IO.Compression.Native/zlib-ng/inffast.h