I'd like us to explore the possibility of consuming zlib-ng, either as one of our zlib implementations, or as the sole one.

https://github.com/zlib-ng/zlib-ng

From the readme:

> The motivation for this fork came after seeing several 3rd party contributions containing new optimizations not getting implemented into the official zlib repository.
> Mark Adler has been maintaining zlib for a very long time, and he has done a great job and hopefully he will continue for a long time yet. The idea of zlib-ng is not to replace zlib, but to co-exist as a drop-in replacement with a lower threshold for code change.

> Features:
> - Zlib compatible API with support for dual-linking
> - Modernized native API based on zlib API for ease of porting
> - Modern C11 syntax and a clean code layout
> - Deflate medium and quick algorithms based on Intels zlib fork
> - Support for CPU intrinsics when available
> - Adler32 implementation using SSSE3, AVX2, Neon, VMX & VSX
> - CRC32-B implementation using PCLMULQDQ & ACLE
> - Hash table implementation using CRC32-C intrinsics on x86 and ARM
> - Slide hash implementations using SSE2, AVX2, Neon, VMX & VSX
> - Compare256/258 implementations using SSE4.2 & AVX2
> - Inflate chunk copying using SSE2, AVX2, Neon & VSX
> - CRC32 implementation using IBM Z vector instructions
> - Support for hardware-accelerated deflate using IBM Z DFLTCC
> - Unaligned memory read/writes and large bit buffer improvements
> - Includes improvements from Cloudflare and Intel forks
> - Configure, CMake, and NMake build system support

Note that Mark Adler is one of the top contributors to that repo: https://github.com/zlib-ng/zlib-ng/graphs/contributors


## Motivation

We currently consume two zlib versions:

- [zlib-intel](https://github.com/dotnet/runtime/blob/4682098d3a435d3f88942ee9e5c6fd426555963a/src/libraries/Native/Windows/System.IO.Compression.Native/CMakeLists.txt#L13-L30): optimized for x86 and x64.
- [zlib](https://github.com/dotnet/runtime/blob/4682098d3a435d3f88942ee9e5c6fd426555963a/src/libraries/Native/Windows/System.IO.Compression.Native/CMakeLists.txt#L31-L43): the original flavor, which is the fallback for all other architectures, like arm and arm64.

The zlib-ng library is optimized for arm and arm64, so we can take advantage of it, and leave the zlib original library as fallback for all other architectures.

## License

The zlib-ng library is provided with the _zlib License_: https://github.com/zlib-ng/zlib-ng/blob/develop/LICENSE.md
It's similar to the _Booster licenses_ from madler/zlib (original zlib flavor): [https://github.com/madler/zlib/contrib/dotzlib/LICENSE_1_0.txt](https://github.com/madler/zlib/blob/cacf7f1d4e3d44d871b605da3b647f07d718623f/contrib/dotzlib/LICENSE_1_0.txt)
And from jtkukunas/zlib (intel): [https://github.com/jtkukunas/zlib/contrib/dotzlib/LICENSE_1_0.txt](https://github.com/jtkukunas/zlib/blob/26df674239f0c8b610822de3be62ee62b9bb9bd3/contrib/dotzlib/LICENSE_1_0.txt)

## Testing

I tested my changes in both my x64 and arm64 machines.
All unit test passed.

## Performance

The perf comparison in my arm64 machine zlib-ng is faster than zlib for most cases:

<details>

  <summary>DeflateStream</summary>

|     Method |        Job |                                                                                                        Toolchain |   level |             file |       Mean |    Error |   StdDev |     Median |        Min |        Max | Ratio | RatioSD |  Gen 0 | Allocated |
|----------- |----------- |----------------------------------------------------------------------------------------------------------------- |-------- |----------------- |-----------:|---------:|---------:|-----------:|-----------:|-----------:|------:|--------:|-------:|----------:|
|   **Compress** | **Job-VVXQFG** |      **\runtime\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe** | **Optimal** | **TestDocument.pdf** | **4,516.9 μs** | **45.25 μs** | **42.33 μs** | **4,497.4 μs** | **4,479.4 μs** | **4,609.3 μs** |  **1.48** |    **0.02** |      **-** |      **8 KB** |
|   Compress | Job-AZGVVU | \runtime_base\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Optimal | TestDocument.pdf | 3,051.0 μs |  9.84 μs |  8.72 μs | 3,047.3 μs | 3,042.8 μs | 3,071.6 μs |  1.00 |    0.00 |      - |      8 KB |
|            |            |                                                                                                                  |         |                  |            |          |          |            |            |            |       |         |        |           |
| Decompress | Job-VVXQFG |      \runtime\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Optimal | TestDocument.pdf |   350.3 μs |  0.96 μs |  0.75 μs |   350.1 μs |   349.6 μs |   352.2 μs |  0.78 |    0.00 | 2.8409 |      8 KB |
| Decompress | Job-AZGVVU | \runtime_base\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Optimal | TestDocument.pdf |   446.4 μs |  0.86 μs |  0.76 μs |   446.4 μs |   445.0 μs |   448.0 μs |  1.00 |    0.00 | 3.5714 |      8 KB |
|            |            |                                                                                                                  |         |                  |            |          |          |            |            |            |       |         |        |           |
|   **Compress** | **Job-VVXQFG** |      **\runtime\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe** | **Optimal** |      **alice29.txt** | **4,599.0 μs** | **11.17 μs** |  **9.90 μs** | **4,599.3 μs** | **4,586.3 μs** | **4,616.1 μs** |  **0.53** |    **0.00** |      **-** |      **8 KB** |
|   Compress | Job-AZGVVU | \runtime_base\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Optimal |      alice29.txt | 8,647.3 μs | 17.93 μs | 15.90 μs | 8,649.4 μs | 8,617.3 μs | 8,676.2 μs |  1.00 |    0.00 |      - |      8 KB |
|            |            |                                                                                                                  |         |                  |            |          |          |            |            |            |       |         |        |           |
| Decompress | Job-VVXQFG |      \runtime\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Optimal |      alice29.txt |   583.3 μs |  3.23 μs |  3.02 μs |   582.4 μs |   578.9 μs |   586.9 μs |  0.98 |    0.01 | 2.3148 |      8 KB |
| Decompress | Job-AZGVVU | \runtime_base\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Optimal |      alice29.txt |   597.5 μs |  4.07 μs |  3.81 μs |   595.8 μs |   593.1 μs |   605.2 μs |  1.00 |    0.00 | 2.4038 |      8 KB |
|            |            |                                                                                                                  |         |                  |            |          |          |            |            |            |       |         |        |           |
|   **Compress** | **Job-VVXQFG** |      **\runtime\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe** | **Optimal** |              **sum** |   **870.6 μs** |  **9.99 μs** |  **9.35 μs** |   **873.6 μs** |   **856.2 μs** |   **889.9 μs** |  **0.44** |    **0.01** | **3.4722** |      **8 KB** |
|   Compress | Job-AZGVVU | \runtime_base\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Optimal |              sum | 1,979.4 μs |  7.90 μs |  7.39 μs | 1,980.1 μs | 1,966.2 μs | 1,993.3 μs |  1.00 |    0.00 |      - |      8 KB |
|            |            |                                                                                                                  |         |                  |            |          |          |            |            |            |       |         |        |           |
| Decompress | Job-VVXQFG |      \runtime\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Optimal |              sum |   134.4 μs |  1.22 μs |  1.02 μs |   134.0 μs |   133.2 μs |   136.3 μs |  0.91 |    0.01 | 3.8043 |      8 KB |
| Decompress | Job-AZGVVU | \runtime_base\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Optimal |              sum |   147.8 μs |  0.92 μs |  0.82 μs |   147.7 μs |   146.1 μs |   149.3 μs |  1.00 |    0.00 | 3.5377 |      8 KB |
|            |            |                                                                                                                  |         |                  |            |          |          |            |            |            |       |         |        |           |
|   **Compress** | **Job-VVXQFG** |      **\runtime\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe** | **Fastest** | **TestDocument.pdf** | **2,628.7 μs** | **11.08 μs** | **10.36 μs** | **2,630.5 μs** | **2,607.2 μs** | **2,645.6 μs** |  **0.91** |    **0.01** |      **-** |      **8 KB** |
|   Compress | Job-AZGVVU | \runtime_base\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Fastest | TestDocument.pdf | 2,901.1 μs | 20.49 μs | 19.17 μs | 2,898.1 μs | 2,874.2 μs | 2,940.5 μs |  1.00 |    0.00 |      - |      8 KB |
|            |            |                                                                                                                  |         |                  |            |          |          |            |            |            |       |         |        |           |
| Decompress | Job-VVXQFG |      \runtime\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Fastest | TestDocument.pdf |   388.8 μs |  0.79 μs |  0.74 μs |   388.8 μs |   387.2 μs |   389.9 μs |  0.86 |    0.00 | 3.1250 |      8 KB |
| Decompress | Job-AZGVVU | \runtime_base\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Fastest | TestDocument.pdf |   452.2 μs |  1.37 μs |  1.28 μs |   451.9 μs |   449.3 μs |   454.7 μs |  1.00 |    0.00 | 3.6765 |      8 KB |
|            |            |                                                                                                                  |         |                  |            |          |          |            |            |            |       |         |        |           |
|   **Compress** | **Job-VVXQFG** |      **\runtime\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe** | **Fastest** |      **alice29.txt** | **2,641.2 μs** | **12.02 μs** | **11.25 μs** | **2,643.9 μs** | **2,620.4 μs** | **2,654.6 μs** |  **1.14** |    **0.01** |      **-** |      **8 KB** |
|   Compress | Job-AZGVVU | \runtime_base\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Fastest |      alice29.txt | 2,308.9 μs | 10.24 μs |  9.07 μs | 2,307.3 μs | 2,297.7 μs | 2,325.3 μs |  1.00 |    0.00 |      - |      8 KB |
|            |            |                                                                                                                  |         |                  |            |          |          |            |            |            |       |         |        |           |
| Decompress | Job-VVXQFG |      \runtime\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Fastest |      alice29.txt |   625.9 μs |  4.90 μs |  4.58 μs |   624.6 μs |   619.8 μs |   632.4 μs |  0.99 |    0.01 | 2.5000 |      8 KB |
| Decompress | Job-AZGVVU | \runtime_base\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Fastest |      alice29.txt |   634.9 μs |  3.48 μs |  3.09 μs |   633.7 μs |   631.1 μs |   641.6 μs |  1.00 |    0.00 | 2.5000 |      8 KB |
|            |            |                                                                                                                  |         |                  |            |          |          |            |            |            |       |         |        |           |
|   **Compress** | **Job-VVXQFG** |      **\runtime\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe** | **Fastest** |              **sum** |   **606.2 μs** |  **3.34 μs** |  **2.96 μs** |   **606.8 μs** |   **598.0 μs** |   **609.4 μs** |  **1.29** |    **0.01** | **2.4038** |      **8 KB** |
|   Compress | Job-AZGVVU | \runtime_base\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Fastest |              sum |   471.3 μs |  3.67 μs |  3.26 μs |   471.1 μs |   466.1 μs |   477.1 μs |  1.00 |    0.00 | 3.7879 |      8 KB |
|            |            |                                                                                                                  |         |                  |            |          |          |            |            |            |       |         |        |           |
| Decompress | Job-VVXQFG |      \runtime\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Fastest |              sum |   136.8 μs |  0.46 μs |  0.43 μs |   136.8 μs |   135.7 μs |   137.3 μs |  0.85 |    0.01 | 3.8043 |      8 KB |
| Decompress | Job-AZGVVU | \runtime_base\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Fastest |              sum |   160.4 μs |  0.92 μs |  0.82 μs |   160.3 μs |   158.9 μs |   162.1 μs |  1.00 |    0.00 | 3.9063 |      8 KB |

</details>

<details>

  <summary>GZipStream</summary>

|     Method |        Job |                                                                                                        Toolchain |   level |             file |       Mean |    Error |   StdDev |     Median |        Min |        Max | Ratio |  Gen 0 | Allocated |
|----------- |----------- |----------------------------------------------------------------------------------------------------------------- |-------- |----------------- |-----------:|---------:|---------:|-----------:|-----------:|-----------:|------:|-------:|----------:|
|   **Compress** | **Job-VVXQFG** |      **\runtime\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe** | **Optimal** | **TestDocument.pdf** | **4,598.6 μs** | **11.21 μs** |  **9.93 μs** | **4,595.6 μs** | **4,586.4 μs** | **4,616.6 μs** |  **1.46** |      **-** |      **8 KB** |
|   Compress | Job-AZGVVU | \runtime_base\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Optimal | TestDocument.pdf | 3,144.2 μs |  6.69 μs |  5.93 μs | 3,144.8 μs | 3,130.6 μs | 3,150.9 μs |  1.00 |      - |      8 KB |
|            |            |                                                                                                                  |         |                  |            |          |          |            |            |            |       |        |           |
| Decompress | Job-VVXQFG |      \runtime\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Optimal | TestDocument.pdf |   452.5 μs |  1.45 μs |  1.28 μs |   452.3 μs |   450.7 μs |   454.9 μs |  0.82 | 3.5714 |      8 KB |
| Decompress | Job-AZGVVU | \runtime_base\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Optimal | TestDocument.pdf |   549.6 μs |  1.08 μs |  0.96 μs |   549.5 μs |   548.1 μs |   551.9 μs |  1.00 | 2.2321 |      8 KB |
|            |            |                                                                                                                  |         |                  |            |          |          |            |            |            |       |        |           |
|   **Compress** | **Job-VVXQFG** |      **\runtime\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe** | **Optimal** |      **alice29.txt** | **4,711.0 μs** |  **4.58 μs** |  **3.82 μs** | **4,710.8 μs** | **4,705.3 μs** | **4,718.9 μs** |  **0.54** |      **-** |      **8 KB** |
|   Compress | Job-AZGVVU | \runtime_base\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Optimal |      alice29.txt | 8,778.4 μs | 18.25 μs | 17.07 μs | 8,781.9 μs | 8,754.0 μs | 8,806.8 μs |  1.00 |      - |      8 KB |
|            |            |                                                                                                                  |         |                  |            |          |          |            |            |            |       |        |           |
| Decompress | Job-VVXQFG |      \runtime\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Optimal |      alice29.txt |   708.1 μs |  4.59 μs |  4.30 μs |   707.4 μs |   703.0 μs |   716.3 μs |  0.98 | 2.8409 |      8 KB |
| Decompress | Job-AZGVVU | \runtime_base\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Optimal |      alice29.txt |   721.1 μs |  4.34 μs |  4.06 μs |   719.5 μs |   716.8 μs |   730.7 μs |  1.00 | 2.8409 |      8 KB |
|            |            |                                                                                                                  |         |                  |            |          |          |            |            |            |       |        |           |
|   **Compress** | **Job-VVXQFG** |      **\runtime\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe** | **Optimal** |              **sum** |   **901.9 μs** | **10.42 μs** |  **9.24 μs** |   **904.1 μs** |   **888.6 μs** |   **915.7 μs** |  **0.45** | **3.4722** |      **8 KB** |
|   Compress | Job-AZGVVU | \runtime_base\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Optimal |              sum | 2,009.7 μs |  6.21 μs |  5.81 μs | 2,009.5 μs | 1,998.0 μs | 2,018.5 μs |  1.00 |      - |      8 KB |
|            |            |                                                                                                                  |         |                  |            |          |          |            |            |            |       |        |           |
| Decompress | Job-VVXQFG |      \runtime\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Optimal |              sum |   168.8 μs |  1.35 μs |  1.26 μs |   168.5 μs |   167.4 μs |   171.3 μs |  0.94 | 4.0323 |      8 KB |
| Decompress | Job-AZGVVU | \runtime_base\artifacts\bin\testhost\net7.0-windows-Release-arm64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe | Optimal |              sum |   180.5 μs |  1.67 μs |  1.56 μs |   180.4 μs |   178.6 μs |   183.8 μs |  1.00 | 3.5920 |      8 KB |
|            |            |                                                                                                                  |         |                  |            |          |          |            |            |            |       |        |           |
…