@carlossanlop 2021-11-21
Need help getting the 3rd commit https://github.com/dotnet/runtime/pull/61883/commits/88b80e530713558842d49a2b32820f9a578309b5 reviewed closely, since I had to make some changes in cmake, `*.c` and `*.h` files to get rid of build failures in arm64.

Also, I am including their license file, the original readme file from the zlib-ng repo. I am imitating one of the other two zlibs, since we also included some non-compilable files.