@jkotas 2021-11-21
> Need help getting the 3rd commit 88b80e5 reviewed closely, since I had to make some changes in cmake, *.c and *.h files to get rid of build failures in arm64.

Some of these changes look like a good candidate for upstreaming into the zlib-ng repo.