@jkotas 2021-11-21
> zlib-intel: optimized for x86 and x64.

Can we delete this copy? The readme for zlib-ng says that it includes all Intel optimizations.