@am11 2021-11-21
This is windows only, right? i.e. on Unix, we still use zlib development package from package management?

Please add a version file similar to: https://github.com/dotnet/runtime/blob/69b5d67d9418d672609aa6e2c418a3d4ae00ad18/src/libraries/Native/AnyOS/brotli-version.txt#L1 and https://github.com/dotnet/runtime/blob/69b5d67d9418d672609aa6e2c418a3d4ae00ad18/src/coreclr/pal/src/libunwind/libunwind-version.txt#L1 to easily keep track of upstream changes.