@nmoinvaz 2021-11-21
It would be better to use zlib-ng on all platforms because then you get the speed benefits on all platforms.