@stephentoub 2021-11-21
> Can we delete this copy? The readme for zlib-ng says that it includes all Intel optimizations.

Taking this a step further,  this is now the third set of zlib files we have in libraries.  Can we delete the other two and just use this one?

> It would be better to use zlib-ng on all platforms

Historically we shied away from this,  preferring to use the zlib inbox; we had to ship it for Windows because Windows didn't have one,  but the Linux distros we do.  However,  at this point,  with the improvements from Intel and CloudFlare, covering both x64 and ARM, with more platform targets varying in what they provide, and with the consistency benefits that come from doing the same thing across platforms,  we should consider changing course and having the sources built into S.IO Compression.Native on all relevant platforms.

> Some of these changes look like a good candidate for upstreaming into the zlib-ng repo.

+1. In addition to the general goodness of contributing back,  for our own maintainability benefits we want to minimize our diff from upstream.