@danmoseley 2021-11-22
It's great to see the improvements in your Arm64 numbers, but some of the regressions are significant. 

For Optimal | TestDocument.pdf, is the output size the same (not sure whether the compression level corresponding to Optimal is part of the spec or not). If not, can we repro this regression without .NET in the picture - perhaps that would be worth an issue opened in the upstream repo. If we become a consumer of it, we should certainly try to help improve it.