@nmoinvaz 2021-11-22
It is worth noting that the compression levels used in zlib-ng are not 1-1 with zlib. This is because zlib-ng uses different/Intel compression algorithms that are not available in zlib.
https://github.com/zlib-ng/zlib-ng/wiki/Deflate-config-comparison

Here are some zlib-ng benchmarks for anybody who is interested:
https://github.com/zlib-ng/zlib-ng/discussions/871
You can also find many benchmarks performed across our PRs.