@mtl1979 2021-11-22
Preferred way of including external CMake projects is using `add_subdirectory()` with `EXCLUDE_FROM_ALL` option... This makes building faster but includes all targets essential for building. Both `source_dir` and `binary_dir` parameters must be specified, as some files are generated by the build system.

Mixing "static" libraries and "dynamic" libraries is discouraged as on most Unix-systems those use different compiler and linker flags and on Windows, the linker can find matching symbols in unexpected places. This also means that order of the libraries (static vs. dynamic/shared) passed to linker is significant.

Build system of zlib-ng currently doesn't support compiling static library as position-independent code (PIC), but the compiler parameters can be passed manually if needed. This works only if no other library in the project (or parent projects) includes or requires zlib-ng.