@carlossanlop 2021-11-23
In PR https://github.com/dotnet/runtime/pull/61958, I was made aware of a zlib implementation located in `src/mono/mono/zlib`.

@lambdageek, @akoeplinger, do you know if it's possible making mono consume the source code related to this PR?