@lambdageek 2021-11-24
> In PR #61958, I was made aware of a zlib implementation located in `src/mono/mono/zlib`.
> 
> @lambdageek, @akoeplinger, do you know if it's possible making mono consume the source code related to this PR?

@carlossanlop I think it should be fine to replace the one in the mono build with zlib-ng.  as far as I know (@akoeplinger please correct me) we just have some old copy of upstream zlib that we update at irregular intervals.