@carlossanlop 2021-11-24
Thanks for confirming, @lambdageek.

My next question: Is `src/libraries/Native` a correct location for code that would be shared by both mono and libraries?