@jkotas 2021-11-24
>  src/libraries/Native a correct location for code that would be shared by both mono and libraries?

It should be src/native (after the other PR merges).