@am11 2021-11-24
@carlossanlop, basically it's https://github.com/am11/runtime/commit/4cbed64ebd7cf09032580d97b3828ed9a052f96e; which will shave off ~11K lines from the repo.